move to walk builder struct + archive template file name
