+++ unpack/Cargo.toml
@@ -26,6 +26,7 @@ serde_regex = "1.1.0"
 regex = { version="1.5.6", default-features = false }
 log = "0.4.0"
 rayon = "1.5.3"
+derive_builder = "0.11.2"
 
 [dev-dependencies]
 insta = { version = "1.21.0", features = ["filters"] }
