+++ unpack/tests/extract_test.rs
@@ -5,7 +5,7 @@ use std::{
     sync::mpsc,
 };
 
-use deep_unpack::{extract_to_folder, extract_to_folder_with_ignores, ArchiveKind, NoWalkList};
+use deep_unpack::{ArchiveKind, DeepWalk, NoWalkList};
 use ignore::WalkBuilder;
 use insta::{assert_debug_snapshot, with_settings};
 use regex::Regex;
@@ -53,16 +53,21 @@ macro_rules! set_snapshot_suffix {
 #[case(1)]
 #[case(2)]
 #[case(3)]
-fn test_can_extract_to_folder(#[case] deep_level: i32) {
-    set_snapshot_suffix!("[level]-[{}]", deep_level);
+fn test_can_extract_with_unpack_level(#[case] unpack_level: u32) {
+    set_snapshot_suffix!("[level]-[{}]", unpack_level);
 
     let destination_folder = get_temp_dir();
     let path = Path::new("tests").join("mocks").join("multiple");
 
     let results = {
-        let mut r = extract_to_folder(&path, &destination_folder.join("dest"), deep_level);
+        let mut r = DeepWalk::new()
+            .folder(format!("{}", path.display()))
+            .unpack_folder(format!("{}", destination_folder.join("dest").display()))
+            .unpack_level(unpack_level)
+            .extract()
+            .unwrap();
 
-        r.sort_by(|a, b| a.path.cmp(&b.path));
+        r.sort_by(|a, b| a.archive_file.cmp(&b.archive_file));
         r
     };
 
@@ -78,7 +83,7 @@ fn test_can_extract_to_folder(#[case] deep_level: i32) {
 }
 
 #[test]
-fn test_can_extract_to_folder_with_custom_ignore() {
+fn test_can_extract_with_ignores() {
     let destination_folder = get_temp_dir();
     let path = Path::new("tests").join("mocks").join("multiple");
 
@@ -89,28 +94,61 @@ fn test_can_extract_to_folder_with_custom_ignore() {
         ],
     };
     let results = {
-        let mut r =
-            extract_to_folder_with_ignores(&path, &destination_folder.join("dest"), 1, &no_walk);
+        let mut r = DeepWalk::new()
+            .folder(format!("{}", path.display()))
+            .unpack_folder(format!("{}", destination_folder.join("dest").display()))
+            .no_walk(no_walk)
+            .extract()
+            .unwrap();
 
-        r.sort_by(|a, b| a.path.cmp(&b.path));
+        r.sort_by(|a, b| a.archive_file.cmp(&b.archive_file));
         r
     };
 
     with_settings!({filters => vec![
-        (r"//*.+/(dest)", "[DYNAMIC-PATH]"),
-        (r"([C]?:\\.+dest\\\\)", "[DYNAMIC-PATH]/"),// for windows
-        (r"\\\\", "/"), // for windows
-    ]}, {
-        assert_debug_snapshot!(get_files_from_folder(&destination_folder));
-        assert_debug_snapshot!(results);
-    });
+            (r"//*.+/(dest)", "[DYNAMIC-PATH]"),
+            (r"([C]?:\\.+dest\\\\)", "[DYNAMIC-PATH]/"),// for windows
+            (r"\\\\", "/"), // for windows
+        ]}, {
+            assert_debug_snapshot!(get_files_from_folder(&
+    destination_folder));         assert_debug_snapshot!(results);
+        });
+    fs::remove_dir_all(destination_folder).unwrap();
+}
+
+#[test]
+fn test_can_extract_with_template() {
+    let destination_folder = get_temp_dir();
+    let path = Path::new("tests").join("mocks").join("multiple");
+
+    let results = {
+        let mut r = DeepWalk::new()
+            .folder(format!("{}", path.display()))
+            .unpack_folder(format!("{}", destination_folder.join("dest").display()))
+            .unpack_level(2 as u32)
+            .extract_template("_PREFIX_${FILENAME}$_SUFFIX_")
+            .extract()
+            .unwrap();
+
+        r.sort_by(|a, b| a.archive_file.cmp(&b.archive_file));
+        r
+    };
+
+    with_settings!({filters => vec![
+            (r"//*.+/(dest)", "[DYNAMIC-PATH]"),
+            (r"([C]?:\\.+dest\\\\)", "[DYNAMIC-PATH]/"),// for windows
+            (r"\\\\", "/"), // for windows
+        ]}, {
+            assert_debug_snapshot!(get_files_from_folder(&
+    destination_folder));         assert_debug_snapshot!(results);
+        });
     fs::remove_dir_all(destination_folder).unwrap();
 }
 
 #[rstest]
 #[case("zip", "archive.zip")]
 #[case("jar", "archive.jar")]
-fn text_can_extract_format(#[case] folder: &str, #[case] file: &str) {
+fn test_can_extract_format(#[case] folder: &str, #[case] file: &str) {
     set_snapshot_suffix!("[{}]-[{}]", folder, file);
 
     let destination_folder = get_temp_dir();
